simplify VT check
