+++ src/libraries/Microsoft.Extensions.Logging.Console/src/ConsoleLoggerProvider.cs
@@ -5,6 +5,7 @@ using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
+using System.Runtime.InteropServices;
 using Microsoft.Extensions.Logging.Configuration;
 using Microsoft.Extensions.Options;
 
@@ -14,7 +15,7 @@ namespace Microsoft.Extensions.Logging.Console
     /// A provider of <see cref="ConsoleLogger"/> instances.
     /// </summary>
     [ProviderAlias("Console")]
-    public partial class ConsoleLoggerProvider : ILoggerProvider, ISupportExternalScope
+    public class ConsoleLoggerProvider : ILoggerProvider, ISupportExternalScope
     {
         private readonly IOptionsMonitor<ConsoleLoggerOptions> _options;
         private readonly ConcurrentDictionary<string, ConsoleLogger> _loggers;
@@ -56,6 +57,22 @@ namespace Microsoft.Extensions.Logging.Console
                 _messageQueue.Console = new AnsiParsingLogConsole();
                 _messageQueue.ErrorConsole = new AnsiParsingLogConsole(stdErr: true);
             }
+        }
+
+        private static bool DoesConsoleSupportAnsi()
+        {
+            if (!RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
+            {
+                return true;
+            }
+            // for Windows, check the console mode
+            var stdOutHandle = Interop.Kernel32.GetStdHandle(Interop.Kernel32.STD_OUTPUT_HANDLE);
+            if (!Interop.Kernel32.GetConsoleMode(stdOutHandle, out int consoleMode))
+            {
+                return false;
+            }
+
+            return (consoleMode & Interop.Kernel32.ENABLE_VIRTUAL_TERMINAL_PROCESSING) == Interop.Kernel32.ENABLE_VIRTUAL_TERMINAL_PROCESSING;
         }
 
         private void SetFormatters(IEnumerable<ConsoleFormatter> formatters = null)
